+++ doc/includes/references.h
@@ -2083,6 +2083,7 @@ typedef struct mjsFlex_ {          // flex specification
   double damping;                  // Rayleigh's damping
   double thickness;                // thickness (2D only)
   int elastic2d;                   // 2D passive forces; 0: none, 1: bending, 2: stretching, 3: both
+  mjtBool snh;                     // experimental 3D stable Neo-Hookean elasticity (mjSpec only)
   int cellcount[3];                // grid cell count for finite cell method
   int order;                       // interpolation order (1: trilinear, 2: quadratic)